ai-agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic-settings>=2.0.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
boto3>=1.34.0
python-docx>=1.1.0
pypdf>=4.0.0
python-multipart>=0.0.6
easyocr>=1.7.1
opencv-python-headless>=4.9.0
numpy>=1.26.0
httpx>=0.26.0
requests>=2.31.0